Simplifies overenrollment calculations.
Instead of trying to keep this value stored in a variable and try
keeping up with the updates whenever it might change, now, the value is
calculated upon request using the array of students that the Course
object has, or using a passed array of all Students in the case of the
overenrollments across all students. Overall, it reduces complexity and
increases confidence, all without slowing down the algorithm incredibly.