Completes the unenrollment process.
The unenrollment process now works! By the end of the process, all
students are enrolled in their maximum number of classes, and
unenrollments occur in the most ideal way. The only tweak that may be
necessary is determining which courses are selected from below the total
min. Right not, it is determined by a ration of the overenrolled
students to the course's min, but this may need to be changed.