Removes checking for underenrolled classes after unenrollment.
Because the unenrollment process decreases section counts as it tries
unenrolling students, by the end of unenrollment, all courses are either
at or above their min. For example; we need to get rid of overenrolled
students in a course. The course loses a section and then unenrolls
students down to the min, meaning it never goes below. If the min goes
to 0, enrolled students can only be above. This means that kicked
students only come from courses above their total max.