Adds the lowest_priority_student() method.
This method will be called a number of times from the driver. It is used
to return a reference to the lowest priority student in a specified
course with a specified number of enrolled courses. It will be used to
determine which students to swap out from classes after they are kicked.