Changes last commit to ignore priority.
There is no reason that a student enrolled in 2 courses should exist
while there are students enrolled in 0 courses. For this reason, the
priority comparison between the current student and marked student has
been removed. This will be added back for the unenrolled students and
the students enrolled in 1 course; then, there is some competition.