Finishes enrolling unenrolled students.
At this point, my algorithm has successfully enrolled each student into
at least one class that is running. For the 600 student test case, only
4 students have gone unenrolled, with 3 due to classes not being offered
at all. The last part of my algorithm needed is the swapping for
students already enrolled in one course who are seeking a second. This
will be done similarly to the unenrolled, using the priority system.
Note that this code will not result in any more unenrolled students; we
are only considering double-enrolled student to be swapped out.