Swaps double-enrolled students based on priorities.
The algorithm is complete! All of the students that have been kicked
from a course have been successfully been either added to a revived
course, added to an open course, swapped in for students taking 2
courses if they have 0, or swapped in for a higher priority. At this
point, all students are enrolled in all courses that they can be. More
testing and checking will follow, as well as refactoring the code.