Modifies restoration of kicked students.
Instead of checking for empty courses first, this has now been
encorporated into the swapping process. This allows for courses that
open up due to swapping to still allow students to be added in later. A
check of these courses was done, and for all classes that are under their
max, there exists no student who can be added in. This was an issue that
existed with the last commit, and is now fixed.