Adds section revival after swapping process.
After all swaps and additions to courses are made after the kicking
process, there is still the chance that another course could be revived
from unenrolled students and single-enrolled students. So, the students
are collected into their single-enrolled and not-enrolled arrays, a new
prefs hash is created, and we look for courses that can be revived. By
this point, all students who can be in a course are in that course, and
if they are not, then either the course is full of higher-priority
students or the course cannot run at all.