Implements course section student division / writing.
The division of course sections is implemented within the method
to_csv(), which takes in the course number and returns an array of
values to be written to the CSV file. Correct course information for
each section of each offered course (excludes those starting with 0
available sections) is now written successfully to the course output
file.