Adds the unenroll_max_from_course() method.
Each of the four major unenroll methods had a similar functionality
which unenrolled the maximum number of students possible from the
course, with this max obviously changing based on how we were
unenrolling. This common code was turned into a parameterized method
that each of the major unenrollment methods now call.